# By darmbrust (25) and others
# Via Maestro (VA) (4) and darmbrust (1)
* 'develop' of https://github.com/Apelon-VA/va-newtons-cradle: (35 commits)
  fix bugs that were preventing the Sememe Provider from working properly when you filter by assemblage
  safer handling of getChronologyTypeForNid, remove an API I previously added that turned out to be unnecessary, fix a nasty bug where incorrect mappings from component nids to concept nids were being made - if someone tried to add a component nid mapping before a sequence identifier had been established for the concept.
  revert some unecessary changes
  just comment changes
  properly handle a concept nid or a component nid in the getConceptNidForNid implementation.
  implement new API call
  just a comment / question to resolve
  add missing shutdown call
  fix sysouts....
  update to API that returns teh task
  move expensive counting code into background task
  disable a slow debug message, add a TODO about the root cause
  reset to snapshot
  [jgitflow-plugin]Updating develop poms back to pre merge state
  [jgitflow-plugin]updating develop poms to master versions to avoid merge conflicts
  [jgitflow-plugin]updating poms for branch'release/1.18' with non-snapshot versions
  [jgitflow-plugin]updating poms for 1.19-SNAPSHOT development
  prep for release
  handle per-concept locking during eConcept import on merge
  just whitespace cleanup
  ...

Conflicts:
	newtons-cradle-integration-tests/pom.xml
	newtons-cradle/src/main/java/gov/vha/isaac/cradle/Cradle.java
	newtons-cradle/src/main/java/gov/vha/isaac/cradle/identifier/IdentifierProvider.java
	newtons-cradle/src/main/java/gov/vha/isaac/cradle/sememe/SememeProvider.java
	newtons-cradle/src/main/java/gov/vha/isaac/cradle/tasks/GenerateIndexes.java

+++ newtons-cradle/src/main/java/gov/vha/isaac/cradle/tasks/GenerateIndexes.java
@@ -13,15 +13,14 @@ import gov.vha.isaac.ochre.api.component.sememe.SememeChronology;
 import gov.vha.isaac.ochre.api.component.sememe.SememeService;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 import javafx.concurrent.Task;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.ihtsdo.otf.lookup.contracts.contracts.ActiveTaskSet;
-import org.ihtsdo.otf.tcc.api.chronicle.ComponentChronicleBI;
-import org.ihtsdo.otf.tcc.model.index.service.IndexStatusListenerBI;
 import org.ihtsdo.otf.tcc.model.cc.refex.RefexMember;
 import org.ihtsdo.otf.tcc.model.cc.refex.RefexService;
+import org.ihtsdo.otf.tcc.model.index.service.IndexStatusListenerBI;
 import org.ihtsdo.otf.tcc.model.index.service.IndexerBI;
 
 /**
@@ -37,15 +36,15 @@ public class GenerateIndexes extends Task<Void> {
     private static final Logger log = LogManager.getLogger();
 
     List<IndexerBI> indexers;
-    int componentCount;
-    AtomicInteger processed = new AtomicInteger(0);
+    long componentCount;
+    AtomicLong processed = new AtomicLong(0);
 
     public GenerateIndexes(Class<?> ... indexersToReindex) {
         updateTitle("Index generation");
         updateProgress(-1, Long.MAX_VALUE); // Indeterminate progress
         if (indexersToReindex == null || indexersToReindex.length == 0)
         {
-        	indexers = LookupService.get().getAllServices(IndexerBI.class);
+            indexers = LookupService.get().getAllServices(IndexerBI.class);
         }
         else
         {
@@ -130,7 +129,7 @@ public class GenerateIndexes extends Task<Void> {
     }
 
     protected void updateProcessedCount() {
-        int processedCount = processed.incrementAndGet();
+        long processedCount = processed.incrementAndGet();
         if (processedCount % 1000 == 0) {
             updateProgress(processedCount, componentCount);
             updateMessage(String.format("Indexed %,d components...", processedCount));
